Linear Regression Progress
Worked on linear regression. Currently, everything theoretically should be working, but in actuality, the y_pred values only form a straight regression line when plotted against X, the input values it was trained on. Currently, models are exhibiting more of a "connect the dots" behavior, where the predicted values exactly match to the input of scattered values.